% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Rcpp_AGS_SIS}
\alias{Rcpp_AGS_SIS}
\title{Adaptive Gibbs Sampler (AGS)
Implementation in C++ of the Adaptive Gibbs Sampler (AGS) for a Generalized Infinite Factor model with Structured Increasing Shrinkage (SIS) prior.}
\usage{
Rcpp_AGS_SIS(
  alpha,
  as,
  a_y,
  a_yp1,
  a_theta,
  Beta,
  bs,
  b0,
  b1,
  burn,
  b_mu,
  b_theta,
  c,
  d,
  eta,
  k,
  kmax,
  kstar,
  Lambda,
  Lambda_star,
  logit,
  mu,
  n,
  nrun,
  out,
  p,
  Phi,
  Plam,
  prec_b,
  prec_mu,
  pred,
  prob,
  ps,
  p_constant,
  q,
  rho,
  sd_b,
  sd_beta,
  sd_mu,
  sp,
  start_adapt,
  thin,
  uu,
  v,
  verbose,
  w,
  W,
  Wnull,
  X
)
}
\arguments{
\item{alpha}{\eqn{Beta(1, \code{alpha})} for pi.}

\item{as, bs}{\eqn{\sigma^{-2} \sim Gamma(\code{as}, \code{bs})}.}

\item{a_y, a_yp1}{Two nxp matrices containing the lower and upper endpoints of the truncated normal distribution.}

\item{a_theta, b_theta}{DA SCRIVERE}

\item{Beta}{A qxk matrix of shrinkage coefficients.}

\item{b0, b1}{Positive constants for the adaptive probability.}

\item{burn}{An integer number of burning iterations (number of iterations to discard).}

\item{b_mu}{A qxc matrix of coefficients.}

\item{c}{An integer number of covariates.}

\item{d}{DA SCRIVERE}

\item{eta}{A nxk matrix of coefficients.}

\item{k}{An integer number of factors.}

\item{kmax}{Maximum number of latent factors.}

\item{kstar}{An integer number of active factors.}

\item{Lambda}{A pxk matrix of coefficients.}

\item{Lambda_star}{A pxk matrix of DA SCRIVERE}

\item{logit}{DA SCRIVERE}

\item{mu}{A pxc matrix of coefficients.}

\item{n}{An integer number of observations.}

\item{nrun}{An integer number of iterations. Default is 100.}

\item{out}{List of variables to save during the Adaptive Gibbs Sampler.}

\item{p}{An integer number of variables.}

\item{Phi}{DA SCRIVERE}

\item{Plam}{DA SCRIVERE}

\item{prec_b}{DA SCRIVERE}

\item{prec_mu}{DA SCRIVERE}

\item{pred}{DA SCRIVERE}

\item{prob}{DA SCRIVERE}

\item{ps}{DA SCRIVERE}

\item{p_constant}{Factor probability constant.}

\item{q}{An integer number of meta-covariates.}

\item{rho}{DA SCRIVERE}

\item{sd_b}{Standard deviation for \eqn{b}.}

\item{sd_beta}{Standard deviation for \eqn{\beta}.}

\item{sd_mu}{Standard deviation for \eqn{\mu}.}

\item{sp}{DA SCRIVERE}

\item{start_adapt}{An integer number of iterations before adaptation.}

\item{thin}{An integer thinning value (number of iterations to skip between saving iterations).}

\item{uu}{DA SCRIVERE}

\item{v}{DA SCRIVERE}

\item{verbose}{Boolean: if \code{true}, print the number of active factors every 10 iterations.}

\item{w}{DA SCRIVERE}

\item{W}{A nxc matrix of covariates.}

\item{Wnull}{Boolean: if \code{true}, the matrix \code{W} of covariates is not considered.}

\item{X}{A pxq matrix of meta-covariates.}
}
\value{
A list.
}
\description{
Adaptive Gibbs Sampler (AGS)
Implementation in C++ of the Adaptive Gibbs Sampler (AGS) for a Generalized Infinite Factor model with Structured Increasing Shrinkage (SIS) prior.
}
\note{
This function uses \code{Rcpp} for computational efficiency.
}
