% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGS_update.R
\name{update_b_mu}
\alias{update_b_mu}
\title{Update b_mu in the Adaptive Gibbs Sampler}
\usage{
update_b_mu(X, prec_mu, prec_b, mu, q, c)
}
\arguments{
\item{X}{DA SCRIVERE}

\item{prec_mu}{DA SCRIVERE}

\item{prec_b}{DA SCRIVERE}

\item{mu}{DA SCRIVERE}

\item{q}{DA SCRIVERE}

\item{c}{DA SCRIVERE}
}
\value{
A qxc matrix.
}
\description{
Update b_mu in the Adaptive Gibbs Sampler
}
