% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGS_update.R
\name{update_b_mu_R}
\alias{update_b_mu_R}
\title{Update b_mu in the Adaptive Gibbs Sampler}
\usage{
update_b_mu_R(X, prec_mu, prec_b, mu, q, c)
}
\arguments{
\item{X}{A pxq matrix.}

\item{prec_mu}{A positive number.}

\item{prec_b}{A positive number.}

\item{mu}{A pxc matrix.}

\item{q}{An integer.}

\item{c}{An integer.}
}
\value{
A qxc matrix.
}
\description{
Update b_mu in the Adaptive Gibbs Sampler
}
