% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGS_update.R
\name{update_beta}
\alias{update_beta}
\title{Update the hth column of beta in the Adaptive Gibbs Sampler}
\usage{
update_beta(h, Xcov, Dt, Bh_1, Phi_L, q)
}
\arguments{
\item{h}{DA SCRIVERE}

\item{Xcov}{DA SCRIVERE}

\item{Dt}{DA SCRIVERE}

\item{Bh_1}{DA SCRIVERE}

\item{Phi_L}{DA SCRIVERE}

\item{q}{DA SCRIVERE}
}
\value{
A qx1 matrix.
}
\description{
Update the hth column of beta in the Adaptive Gibbs Sampler
}
