% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_bmu}
\alias{update_bmu}
\title{Update b_mu in the Adaptive Gibbs Sampler}
\usage{
update_bmu(X, prec_mu, prec_b, mu, q, c)
}
\arguments{
\item{X}{A pxq matrix.}

\item{prec_mu}{A positive number.}

\item{prec_b}{A positive number.}

\item{mu}{A pxc matrix.}

\item{q}{An integer.}

\item{c}{An integer.}
}
\value{
A qxc matrix.
}
\description{
Update b_mu in the Adaptive Gibbs Sampler
}
\note{
This function uses \code{Rcpp} for computational efficiency.
}
