% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGS_update.R
\name{update_d_R}
\alias{update_d_R}
\title{Update the hth element of d in the Adaptive Gibbs Sampler}
\usage{
update_d_R(h, Phi, p, n, rho, etalambdastar, Z, sdy, k, w)
}
\arguments{
\item{h}{DA SCRIVERE}

\item{Phi}{DA SCRIVERE}

\item{p}{DA SCRIVERE}

\item{n}{DA SCRIVERE}

\item{rho}{DA SCRIVERE}

\item{etalambdastar}{DA SCRIVERE}

\item{Z}{DA SCRIVERE}

\item{sdy}{DA SCRIVERE}

\item{k}{DA SCRIVERE}

\item{w}{DA SCRIVERE}
}
\value{
An integer in 1, ..., k.
}
\description{
Update the hth element of d in the Adaptive Gibbs Sampler
}
