% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_eta}
\alias{update_eta}
\title{Update eta in the Adaptive Gibbs Sampler}
\usage{
update_eta(Lambda, ps, k, Z, n)
}
\arguments{
\item{Lambda}{A pxk matrix.}

\item{ps}{A p-dimensional vector.}

\item{k}{An integer.}

\item{Z}{A nxp matrix.}

\item{n}{An integer.}
}
\value{
A nxk matrix.
}
\description{
Update eta in the Adaptive Gibbs Sampler
}
\note{
This function uses \code{Rcpp} for computational efficiency.
}
