% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_Lambda_star}
\alias{update_Lambda_star}
\title{Update the jth row of Lambda_star in the Adaptive Gibbs Sampler}
\usage{
update_Lambda_star(j, etarho, Phi, Plam, ps, Z, k)
}
\arguments{
\item{j}{An integer.}

\item{etarho}{A kxn matrix.}

\item{Phi}{A pxk matrix.}

\item{Plam}{A kxk matrix.}

\item{ps}{A p-dimensional vector;}

\item{Z}{A nxp matrix.}

\item{k}{An integer.}
}
\value{
A 1xk matrix.
}
\description{
Update the jth row of Lambda_star in the Adaptive Gibbs Sampler
}
\note{
This function uses \code{Rcpp} for computational efficiency.
}
