% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_mu}
\alias{update_mu}
\title{Update the hth row of mu in the Adaptive Gibbs Sampler}
\usage{
update_mu(j, Qbet, W, Z_res, ps, b_mu, Xcov, c)
}
\arguments{
\item{j}{An integer.}

\item{Qbet}{A cxc matrix.}

\item{W}{A nxc matrix.}

\item{Z_res}{A nxp matrix.}

\item{ps}{A p-dimensional vector.}

\item{b_mu}{A cxq matrix.}

\item{Xcov}{A pxq matrix.}

\item{c}{An integer.}
}
\value{
A 1xc matrix.
}
\description{
Update the hth row of mu in the Adaptive Gibbs Sampler
}
\note{
This function uses \code{Rcpp} for computational efficiency.
}
