% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGS_update.R
\name{update_Phi}
\alias{update_Phi}
\title{Update Phi in the Adaptive Gibbs Sampler}
\usage{
update_Phi(rho, logit, p_constant, p, n, eta, Lambda_star, Phi, Z, sdy)
}
\arguments{
\item{rho}{DA SCRIVERE}

\item{logit}{DA SCRIVERE}

\item{p_constant}{DA SCRIVERE}

\item{p}{DA SCRIVERE}

\item{n}{DA SCRIVERE}

\item{eta}{DA SCRIVERE}

\item{Lambda_star}{DA SCRIVERE}

\item{Phi}{DA SCRIVERE}

\item{Z}{DA SCRIVERE}

\item{sdy}{DA SCRIVERE}
}
\value{
A pxk matrix.
}
\description{
Update Phi in the Adaptive Gibbs Sampler
}
