% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGS.R
\name{AGS_SIS}
\alias{AGS_SIS}
\title{AGS for GIF models with SIS prior (C++)}
\usage{
AGS_SIS(
  Y,
  X = NULL,
  W = NULL,
  seed = 28,
  stdx = TRUE,
  stdw = TRUE,
  WFormula = formula("~ ."),
  XFormula = formula("~ ."),
  kinit = NULL,
  kmax = NULL,
  kval = 6,
  nrun = 100,
  burn = round(nrun/4),
  thin = 1,
  start_adapt = 50,
  b0 = 1,
  b1 = 5 * 10^(-4),
  sd_b = 1,
  sd_mu = 1,
  sd_beta = 1,
  a_theta = 1,
  b_theta = 1,
  as = 1,
  bs = 1,
  p_constant = NULL,
  alpha,
  y_max = Inf,
  output = "all",
  verbose = TRUE
)
}
\arguments{
\item{Y}{A nxp matrix of counts.}

\item{X}{A pxq matrix of meta-covariates; the variables must be numeric or factors.}

\item{W}{A nxc matrix of covariates; the variables must be numeric or factors.}

\item{seed}{Seed. Default is 28.}

\item{stdx}{Boolean: if TRUE, numeric meta-covariates are standardized.}

\item{stdw}{Boolean: if TRUE, numeric covariates are standardized.}

\item{WFormula}{Formula specifying  the covariates inserted in the model.}

\item{XFormula}{Formula specifying  the meta-covariates inserted in the model.}

\item{kinit}{An integer minimun number of latent factors. Default is \code{min(floor(log(p)*kval),p)}.}

\item{kmax}{Maximum number of latent factors. Default is \code{p+1}.}

\item{kval}{An integer initial number of latent factors. Default is 6.}

\item{nrun}{An integer number of iterations. Default is 100.}

\item{burn}{An integer number of burning iterations (number of iterations to discard). Default is \code{round(nrun/4)}.}

\item{thin}{An integer thinning value (number of iterations to skip between saving iterations). Default is 1.}

\item{start_adapt}{An integer number of iterations before adaptation. Default is 50.}

\item{b0, b1}{Positive constants for the adaptive probability. Default is \code{c(1,5*10^(-4))}.}

\item{sd_b}{Standard deviation for \eqn{b}. Default is 1.}

\item{sd_mu}{Standard deviation for \eqn{\mu}. Default is 1.}

\item{sd_beta}{Standard deviation for \eqn{\beta}. Default is 1.}

\item{a_theta, b_theta}{Default is \code{c(1,1)}.}

\item{as, bs}{\eqn{\sigma^{-2} \sim Gamma(\code{as}, \code{bs})}. Default is \code{c(1,1)}.}

\item{p_constant}{Factor probability constant. Default is \code{10*exp(1)*log(p)/p}.}

\item{alpha}{\eqn{Beta(1, \code{alpha})} for pi.}

\item{y_max}{A fixed and known upper bound for the values in \code{Y}. Default is \code{Inf}.}

\item{output}{Default is \code{"all"}.}

\item{verbose}{If TRUE, print the number of active factors every 10 iterations. Default is \code{TRUE}.}
}
\value{
A list.
}
\description{
Implementation in C++ of the Adaptive Gibbs Sampler (AGS) for a Generalized Infinite Factor model with Structured Increasing Shrinkage (SIS) prior.
}
\seealso{
\code{\link{update_bmu}}, \code{\link{update_mu}}, \code{\link{update_eta}}, \code{\link{update_Lambda_star}}, \code{\link{update_d}}, \code{\link{update_Phi}}
}
