% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_beta}
\alias{update_beta}
\title{Update the hth column of beta in the Adaptive Gibbs Sampler}
\usage{
update_beta(h, Xcov, Dt, Bh_1, Phi_L, q)
}
\arguments{
\item{h}{An integer.}

\item{Xcov}{A pxq matrix.}

\item{Dt}{A pxk matrix.}

\item{Bh_1}{A qxq matrix.}

\item{Phi_L}{A pxk matrix.}

\item{q}{An integer.}
}
\value{
A qx1 matrix.
}
\description{
Update the hth column of beta in the Adaptive Gibbs Sampler
}
\note{
This function uses \code{Rcpp} for computational efficiency.
}
