% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_d}
\alias{update_d}
\title{Update the hth element of d in the Adaptive Gibbs Sampler}
\usage{
update_d(h, Phi, p, n, rho, eta, lambdastar, Z, sdy, k, w)
}
\arguments{
\item{h}{An integer.}

\item{Phi}{A pxk matrix.}

\item{p}{An integer.}

\item{n}{An integer.}

\item{rho}{A k-dimensional vector.}

\item{eta}{A nxk matrix.}

\item{lambdastar}{A pxk matrix.}

\item{Z}{A nxp matrix.}

\item{sdy}{A nxp matrix.}

\item{k}{An integer.}

\item{w}{A k-dimensional vector.}
}
\value{
An integer in 1, ..., k.
}
\description{
Update the hth element of d in the Adaptive Gibbs Sampler
}
\note{
This function uses \code{Rcpp} for computational efficiency.
}
