% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGS_update.R
\name{update_eta}
\alias{update_eta}
\title{Update eta in the Adaptive Gibbs Sampler}
\usage{
update_eta(Lambda, ps, k, Z, n)
}
\arguments{
\item{Lambda}{A pxk matrix.}

\item{ps}{DA SCRIVERE}

\item{k}{Number of columns of eta.}

\item{Z}{A nxp matrix.}

\item{n}{Number of rows of eta.}
}
\value{
A nxk matrix.
}
\description{
Update eta in the Adaptive Gibbs Sampler
}
