% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGS_update.R
\name{update_Lambda_star}
\alias{update_Lambda_star}
\title{Update the jth row of Lambda_star in the Adaptive Gibbs Sampler}
\usage{
update_Lambda_star(j, etarho, Phi, Plam, ps, Z, k)
}
\arguments{
\item{j}{DA SCRIVERE}

\item{etarho}{DA SCRIVERE}

\item{Phi}{DA SCRIVERE}

\item{Plam}{DA SCRIVERE}

\item{ps}{DA SCRIVERE}

\item{Z}{DA SCRIVERE}

\item{k}{DA SCRIVERE}
}
\value{
A 1xk matrix.
}
\description{
Update the jth row of Lambda_star in the Adaptive Gibbs Sampler
}
