% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AGS_update.R
\name{update_mu}
\alias{update_mu}
\title{Update the hth row of mu in the Adaptive Gibbs Sampler}
\usage{
update_mu(j, Qbet, W, Z_res, ps, b_mu, Xcov, c)
}
\arguments{
\item{j}{DA SCRIVERE}

\item{Qbet}{DA SCRIVERE}

\item{Z_res}{DA SCRIVERE}

\item{ps}{DA SCRIVERE}

\item{b_mu}{DA SCRIVERE}

\item{Xcov}{DA SCRIVERE}

\item{c}{DA SCRIVERE}
}
\value{
A 1xc matrix.
}
\description{
Update the hth row of mu in the Adaptive Gibbs Sampler
}
