% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{update_Phi}
\alias{update_Phi}
\title{Update Phi in the Adaptive Gibbs Sampler}
\usage{
update_Phi(rho, logit, p_constant, p, n, eta, lambdastar, Phi, Z, sdy, k)
}
\arguments{
\item{rho}{A k-dimensional vector.}

\item{logit}{A pxk matrix.}

\item{p_constant}{A number in (0,1).}

\item{p}{An integer.}

\item{n}{An integer.}

\item{eta}{A nxk matrix.}

\item{lambdastar}{A pxk matrix.}

\item{Phi}{A pxk matrix.}

\item{Z}{A nxp matrix.}

\item{sdy}{A nxp matrix.}

\item{k}{An integer.}
}
\value{
A pxk matrix.
}
\description{
Update Phi in the Adaptive Gibbs Sampler
}
\note{
This function uses \code{Rcpp} for computational efficiency.
}
